+++ src/Common/buildEntityConfig.ts
@@ -2,7 +2,7 @@ import { StringsKey, getString } from '@utils/getString';
 
 export const buildEntityConfig = (
   key: StringsKey,
-  additionalConfig?: string | { category?: string; icon?: string }
+  additionalConfig?: string | { category?: string; icon?: string; deviceClass?: string }
 ) => {
   if (typeof additionalConfig === 'string') additionalConfig = { category: additionalConfig };
   return {